Hoist remote client and auth checker into runmode
The remote API-client constructor and the credential checker were
duplicated across packages: `sandboxcmd.getRemoteClient` /
`sandboxcmd.defaultAuthChecker` and the `service` command's
`newRemoteClient` / `remoteAuthChecker` (plus `getSecretsClient`) were
all copies of the same two snippets.

Add `runmode.NewRemoteClient` and `runmode.DefaultAuthChecker` as the
single source of truth and route every caller through them. The env-var
branch stays in `DefaultAuthChecker` so it remains a complete standalone
credential check (it's covered by a dedicated test), even though
`RequireAuth` also short-circuits on `AMIKA_API_KEY`.

No behavior change: `getRemoteClient` never returned an error and
ignored its `target` (validated separately by `getRemoteTarget`), so
call sites drop the dead error handling. Moves the auth-checker tests to
the runmode package alongside the code.
